Move argument parsing data to separate module.
